Step 1 to split out password, mail, create-user
