+++ src/server.js
@@ -1,4 +1,4 @@
-// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
+P// Copyright (C) 2012-present, The Authors. This program is free software: you can redistribute it and/or  modify it under the terms of the GNU Affero General Public License, version 3, as published by the Free Software Foundation. This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details. You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 "use strict";
 
@@ -62,6 +62,16 @@ const pgQueryP_metered = pg.queryP_metered;
 const pgQueryP_metered_readOnly = pg.queryP_metered_readOnly;
 const pgQueryP_readOnly = pg.queryP_readOnly;
 const pgQueryP_readOnly_wRetryIfEmpty = pg.queryP_readOnly_wRetryIfEmpty;
+
+const createUser = require('./auth/create-user');
+const doSendVerification = createUser.doSendVerification;
+const generateToken = createUser.generateToken;
+const generateTokenP = createUser.generateTokenP;
+const generateAndRegisterZinvite = createUser.generateAndRegisterZinvite;
+
+// TODO: Maybe able to remove
+const generateHashedPassword = require('./auth/password').generateHashedPassword;
+const checkPassword = require('./auth/password').checkPassword;
 // End of re-import
 
 i18n.configure({
@@ -951,23 +961,8 @@ String.prototype.hashCode = function() {
   return hash;
 };
 
-function fail(res, httpCode, clientVisibleErrorString, err) {
-  emitTheFailure(res, httpCode, "polis_err", clientVisibleErrorString, err);
-  yell(clientVisibleErrorString);
-}
-
-function userFail(res, httpCode, clientVisibleErrorString, err) {
-  emitTheFailure(res, httpCode, "polis_user_err", clientVisibleErrorString, err);
-}
-
-function emitTheFailure(res, httpCode, extraErrorCodeForLogs, clientVisibleErrorString, err) {
-  console.error(clientVisibleErrorString, extraErrorCodeForLogs, err);
-  if (err && err.stack) {
-    console.error(err.stack);
-  }
-  res.writeHead(httpCode || 500);
-  res.end(clientVisibleErrorString);
-}
+const fail = log.fail;
+const userFail = log.userFail;
 
 function isEmail(s) {
   return typeof s === "string" && s.length < 999 && s.indexOf("@") > 0;
@@ -1607,20 +1602,6 @@ function initializePolisHelpers() {
     } else {
       return req.cookies[COOKIES.PERMANENT_COOKIE];
     }
-  }
-
-  function generateHashedPassword(password, callback) {
-    bcrypt.genSalt(12, function(errSalt, salt) {
-      if (errSalt) {
-        return callback("polis_err_salt");
-      }
-      bcrypt.hash(password, salt, function(errHash, hashedPassword) {
-        if (errHash) {
-          return callback("polis_err_hash");
-        }
-        callback(null, hashedPassword);
-      });
-    });
   }
 
   let pidCache = new LruCache({
@@ -3609,173 +3590,6 @@ Feel free to reply to this email if you need help.`;
     });
   }
 
-  function generateTokenP(len, pseudoRandomOk) {
-    return new Promise(function(resolve, reject) {
-      generateToken(len, pseudoRandomOk, function(err, token) {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(token);
-        }
-      });
-    });
-  }
-
-  function generateToken(len, pseudoRandomOk, callback) {
-    // TODO store up a buffer of random bytes sampled at random times to reduce predictability. (or see if crypto module does this for us)
-    // TODO if you want more readable tokens, see ReadableIds
-    let gen;
-    if (pseudoRandomOk) {
-      gen = crypto.pseudoRandomBytes;
-    } else {
-      gen = crypto.randomBytes;
-    }
-    gen(len, function(err, buf) {
-      if (err) {
-        return callback(err);
-      }
-
-      let prettyToken = buf.toString('base64')
-        .replace(/\//g, 'A').replace(/\+/g, 'B') // replace url-unsafe tokens (ends up not being a proper encoding since it maps onto A and B. Don't want to use any punctuation.)
-        .replace(/l/g, 'C') // looks like '1'
-        .replace(/L/g, 'D') // looks like '1'
-        .replace(/o/g, 'E') // looks like 0
-        .replace(/O/g, 'F') // looks lke 0
-        .replace(/1/g, 'G') // looks like 'l'
-        .replace(/0/g, 'H') // looks like 'O'
-        .replace(/I/g, 'J') // looks like 'l'
-        .replace(/g/g, 'K') // looks like 'g'
-        .replace(/G/g, 'M') // looks like 'g'
-        .replace(/q/g, 'N') // looks like 'q'
-        .replace(/Q/g, 'R') // looks like 'q'
-      ;
-      // replace first character with a number between 2 and 9 (avoiding 0 and 1 since they look like l and O)
-      prettyToken = _.random(2, 9) + prettyToken.slice(1);
-      prettyToken = prettyToken.toLowerCase();
-      prettyToken = prettyToken.slice(0, len); // in case it's too long
-
-      callback(0, prettyToken);
-    });
-  }
-
-  // function generateApiKeyForUser(uid, optionalPrefix) {
-  //   let parts = ["pkey"];
-  //   let len = 32;
-  //   if (!_.isUndefined(optionalPrefix)) {
-  //     parts.push(optionalPrefix);
-  //   }
-  //   len -= parts[0].length;
-  //   len -= (parts.length - 1); // the underscores
-  //   parts.forEach(function(part) {
-  //     len -= part.length;
-  //   });
-  //   return generateTokenP(len, false).then(function(token) {
-  //     parts.push(token);
-  //     let apikey = parts.join("_");
-  //     return apikey;
-  //   });
-  // }
-
-  // function addApiKeyForUser(uid, optionalPrefix) {
-  //   return generateApiKeyForUser(uid, optionalPrefix).then(function(apikey) {
-  //     return pgQueryP("insert into apikeysndvweifu (uid, apikey)  VALUES ($1, $2);", [uid, apikey]);
-  //   });
-  // }
-
-
-  // function getApiKeysTruncated(uid) {
-  //   return pgQueryP_readOnly("select * from apikeysndvweifu WHERE uid = ($1);", [uid]).then(function(rows) {
-  //     if (!rows || !rows.length) {
-  //       return [];
-  //     }
-  //     return rows.map(function(row) {
-  //       return {
-  //         apikeyTruncated: row.apikey.slice(0, 10) + "...",
-  //         created: row.created,
-  //       };
-  //     });
-  //   });
-  // }
-
-  // function createApiKey(uid) {
-  //   return generateTokenP(17, false).then(function(token) {
-  //     let apikey = "pkey_" + token;
-  //     return pgQueryP("insert into apikeysndvweifu (uid, apikey) values ($1, $2) returning *;", [uid, apikey]).then(function(row) {
-  //       return {
-  //         apikey: apikey,
-  //         created: row.created,
-  //       };
-  //     });
-  //   });
-  // }
-
-  // function deleteApiKey(uid, apikeyTruncated) {
-  //   // strip trailing "..."
-  //   apikeyTruncated = apikeyTruncated.slice(0, apikeyTruncated.indexOf("."));
-  //   // basic sanitizing - replace unexpected characters with x's.
-  //   apikeyTruncated = apikeyTruncated.replace(/[^a-zA-Z0-9_]/g, 'x');
-  //   return pgQueryP("delete from apikeysndvweifu where uid = ($1) and apikey ~ '^" + apikeyTruncated + "';", [uid]);
-  // }
-
-
-
-  // function addApiKeyForUsersBulk(uids, optionalPrefix) {
-  //     let promises = uids.map(function(uid) {
-  //         return generateApiKeyForUser(uid, optionalPrefix);
-  //     });
-  //     return Promise.all(promises).then(function(apikeys) {
-  //         let query = "insert into apikeysndvweifu (uid, apikey)  VALUES ";
-  //         let pairs = [];
-  //         for (var i = 0; i < uids.length; i++) {
-  //             let uid = uids[i];
-  //             let apikey = apikeys[i];
-  //             pairs.push("(" + uid + ', \'' + apikey + '\')');
-  //         }
-  //         query += pairs.join(',');
-  //         query += 'returning uid;';
-  //         return pgQueryP(query, []);
-  //     });
-  // }
-
-  // let uidsX = [];
-  // for (var i = 200200; i < 300000; i++) {
-  //     uidsX.push(i);
-  // }
-  // addApiKeyForUsersBulk(uidsX, "test23").then(function(uids) {
-  //     console.log("hihihihi", uids.length);
-  //     setTimeout(function() { process.exit();}, 3000);
-  // });
-
-  // // let time1 = Date.now();
-  // createDummyUsersBatch(3 * 1000).then(function(uids) {
-  //         // let time2 = Date.now();
-  //         // let dt = time2 - time1;
-  //         // console.log("time foo" , dt);
-  //         // console.dir(uids);
-  //         uids.forEach(function(uid) {
-  //             console.log("hihihihi", uid);
-  //         });
-  //         process.exit(0);
-
-  // }).catch(function(err) {
-  //     console.error("errorfooooo");
-  //     console.error(err);
-  // });
-
-
-  function generateAndRegisterZinvite(zid, generateShort) {
-    let len = 10;
-    if (generateShort) {
-      len = 6;
-    }
-    return generateTokenP(len, false).then(function(zinvite) {
-      return pgQueryP('INSERT INTO zinvites (zid, zinvite, created) VALUES ($1, $2, default);', [zid, zinvite]).then(function(rows) {
-        return zinvite;
-      });
-    });
-  }
-
-
 
   function handle_POST_zinvites(req, res) {
     let generateShortUrl = req.p.short_url;
@@ -4528,16 +4342,6 @@ ${serverName}/pwreset/${pwresettoken}
       body);
   }
 
-  function sendVerificaionEmail(req, email, einvite) {
-    let serverName = getServerNameWithProtocol(req);
-    let body = i18n.__("PolisVerification", serverName, einvite);
-    return sendTextEmail(
-      POLIS_FROM_ADDRESS,
-      email,
-      i18n.__("Polis verification"),
-      body);
-  }
-
   function isEmailVerified(email) {
     return pgQueryP("select * from email_validations where email = ($1);", [email]).then(function(rows) {
       return rows.length > 0;
@@ -4837,26 +4641,6 @@ Email verified! You can close this tab or hit the back button.
       if (!rows || !rows.length) {
         return pgQueryP("insert into lti_context_memberships (uid, lti_context_id, tool_consumer_instance_guid) values ($1, $2, $3);", [uid, lti_context_id, tool_consumer_instance_guid]);
       }
-    });
-  }
-
-  function checkPassword(uid, password) {
-    return pgQueryP_readOnly_wRetryIfEmpty("select pwhash from jianiuevyew where uid = ($1);", [uid]).then(function(rows) {
-      if (!rows || !rows.length) {
-        return null;
-      } else if (!rows[0].pwhash) {
-        return void 0;
-      }
-      let hashedPassword = rows[0].pwhash;
-      return new Promise(function(resolve, reject) {
-        bcrypt.compare(password, hashedPassword, function(errCompare, result) {
-          if (errCompare) {
-            reject(errCompare);
-          } else {
-            resolve(result ? "ok" : 0);
-          }
-        });
-      });
     });
   }
 
@@ -6421,187 +6205,7 @@ Email verified! You can close this tab or hit the back button.
   } // end do_handle_POST_auth_facebook
 
   function handle_POST_auth_new(req, res) {
-    let hname = req.p.hname;
-    let password = req.p.password;
-    let password2 = req.p.password2; // for verification
-    let email = req.p.email;
-    let oinvite = req.p.oinvite;
-    let zinvite = req.p.zinvite;
-    let referrer = req.cookies[COOKIES.REFERRER];
-    let organization = req.p.organization;
-    let gatekeeperTosPrivacy = req.p.gatekeeperTosPrivacy;
-    let lti_user_id = req.p.lti_user_id;
-    let lti_user_image = req.p.lti_user_image;
-    let lti_context_id = req.p.lti_context_id;
-    let tool_consumer_instance_guid = req.p.tool_consumer_instance_guid;
-    let afterJoinRedirectUrl = req.p.afterJoinRedirectUrl;
-
-    let site_id = void 0;
-    if (req.p.encodedParams) {
-      let decodedParams = decodeParams(req.p.encodedParams);
-      if (decodedParams.site_id) {
-        // NOTE: we could have just allowed site_id to be passed as a normal param, but then we'd need to think about securing that with some other token sooner.
-        // I think we can get by with this obscure scheme for a bit.
-        // TODO_SECURITY add the extra token associated with the site_id owner.
-        site_id = decodedParams.site_id;
-      }
-    }
-
-    let shouldAddToIntercom = req.p.owner;
-    if (req.p.lti_user_id) {
-      shouldAddToIntercom = false;
-    }
-
-    if (password2 && (password !== password2)) {
-      fail(res, 400, "Passwords do not match.");
-      return;
-    }
-    if (!gatekeeperTosPrivacy) {
-      fail(res, 400, "polis_err_reg_need_tos");
-      return;
-    }
-    if (!email) {
-      fail(res, 400, "polis_err_reg_need_email");
-      return;
-    }
-    if (!hname) {
-      fail(res, 400, "polis_err_reg_need_name");
-      return;
-    }
-    if (!password) {
-      fail(res, 400, "polis_err_reg_password");
-      return;
-    }
-    if (password.length < 6) {
-      fail(res, 400, "polis_err_reg_password_too_short");
-      return;
-    }
-    if (!_.contains(email, "@") || email.length < 3) {
-      fail(res, 400, "polis_err_reg_bad_email");
-      return;
-    }
-
-    pgQueryP("SELECT * FROM users WHERE email = ($1)", [email]).then(function(rows) {
-
-      if (rows.length > 0) {
-        fail(res, 403, "polis_err_reg_user_with_that_email_exists");
-        return;
-      }
-
-      generateHashedPassword(password, function(err, hashedPassword) {
-        if (err) {
-          fail(res, 500, "polis_err_generating_hash", err);
-          return;
-        }
-        let query = "insert into users " +
-          "(email, hname, zinvite, oinvite, is_owner" + (site_id ? ", site_id" : "") + ") VALUES " + // TODO use sql query builder
-          "($1, $2, $3, $4, $5" + (site_id ? ", $6" : "") + ") " + // TODO use sql query builder
-          "returning uid;";
-        let vals =
-          [email, hname, zinvite || null, oinvite || null, true];
-        if (site_id) {
-          vals.push(site_id); // TODO use sql query builder
-        }
-
-        doSendVerification(req, email);
-
-        pgQuery(query, vals, function(err, result) {
-          if (err) {
-            winston.log("info", err);
-            fail(res, 500, "polis_err_reg_failed_to_add_user_record", err);
-            return;
-          }
-          let uid = result && result.rows && result.rows[0] && result.rows[0].uid;
-
-          pgQuery("insert into jianiuevyew (uid, pwhash) values ($1, $2);", [uid, hashedPassword], function(err, results) {
-            if (err) {
-              winston.log("info", err);
-              fail(res, 500, "polis_err_reg_failed_to_add_user_record", err);
-              return;
-            }
-
-
-            startSession(uid, function(err, token) {
-              if (err) {
-                fail(res, 500, "polis_err_reg_failed_to_start_session", err);
-                return;
-              }
-              addCookies(req, res, token, uid).then(function() {
-
-                let ltiUserPromise = lti_user_id ?
-                  addLtiUserifNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) :
-                  Promise.resolve();
-                let ltiContextMembershipPromise = lti_context_id ?
-                  addLtiContextMembership(uid, lti_context_id, tool_consumer_instance_guid) :
-                  Promise.resolve();
-                Promise.all([ltiUserPromise, ltiContextMembershipPromise]).then(function() {
-                  if (lti_user_id) {
-                    if (afterJoinRedirectUrl) {
-                      res.redirect(afterJoinRedirectUrl);
-                    } else {
-                      renderLtiLinkageSuccessPage(req, res, {
-                        // may include token here too
-                        context_id: lti_context_id,
-                        uid: uid,
-                        hname: hname,
-                        email: email,
-                      });
-                    }
-                  } else {
-                    res.json({
-                      uid: uid,
-                      hname: hname,
-                      email: email,
-                      // token: token
-                    });
-                  }
-                }).catch(function(err) {
-                  fail(res, 500, "polis_err_creating_user_associating_with_lti_user", err);
-                });
-
-                if (shouldAddToIntercom) {
-                  let params = {
-                    "email": email,
-                    "name": hname,
-                    "user_id": uid,
-                  };
-                  let customData = {};
-                  if (referrer) {
-                    customData.referrer = referrer;
-                  }
-                  if (organization) {
-                    customData.org = organization;
-                  }
-                  customData.uid = uid;
-                  if (_.keys(customData).length) {
-                    params.custom_data = customData;
-                  }
-                  // Do not know what is intercom, just skip
-                  if (intercom.createUser) {
-                      intercom.createUser(params, function (err, res) {
-                          if (err) {
-                              winston.log("info", err);
-                              console.error("polis_err_intercom_create_user_fail");
-                              winston.log("info", params);
-                              yell("polis_err_intercom_create_user_fail");
-                              return;
-                          }
-                      });
-                  }
-                }
-              }, function(err) {
-                fail(res, 500, "polis_err_adding_cookies", err);
-              }).catch(function(err) {
-                fail(res, 500, "polis_err_adding_user", err);
-              });
-            }); // end startSession
-          }); // end insert pwhash
-        }); // end insert user
-      }); // end generateHashedPassword
-
-    }, function(err) {
-      fail(res, 500, "polis_err_reg_checking_existing_users", err);
-    });
+    require('./auth/create-user').createUser(req, res, COOKIES);
   } // end /api/v3/auth/new
 
 
@@ -10288,20 +9892,6 @@ Email verified! You can close this tab or hit the back button.
   }
 
 
-  function decodeParams(encodedStringifiedJson) {
-    if (!encodedStringifiedJson.match(/^\/?ep1_/)) {
-      throw new Error("wrong encoded params prefix");
-    }
-    if (encodedStringifiedJson[0] === "/") {
-      encodedStringifiedJson = encodedStringifiedJson.slice(5);
-    } else {
-      encodedStringifiedJson = encodedStringifiedJson.slice(4);
-    }
-    let stringifiedJson = hexToStr(encodedStringifiedJson);
-    let o = JSON.parse(stringifiedJson);
-    return o;
-  }
-
   function encodeParams(o) {
     let stringifiedJson = JSON.stringify(o);
     let encoded = "ep1_" + strToHex(stringifiedJson);
@@ -12499,16 +12089,6 @@ Thanks for using Polis!
       });
     });
   }
-
-  function doSendVerification(req, email) {
-    return generateTokenP(30, false).then(function(einvite) {
-      return pgQueryP("insert into einvites (email, einvite) values ($1, $2);", [email, einvite]).then(function(rows) {
-        return sendVerificaionEmail(req, email, einvite);
-      });
-    });
-  }
-
-
 
   function handle_GET_slack_login(req, res) {
